telephone number type changed from int to long

+++ Form1.cs
@@ -27,7 +27,7 @@ namespace Contacts
 
         private void saveBtn_Click(object sender, EventArgs e)
         {
-            TelephoneDirectory.contactList.Add(new Person(nameBox.Text, addressBox.Text, int.Parse(ageBox.Text), int.Parse(mobileBox.Text), typeComboBox.Text));
+            TelephoneDirectory.contactList.Add(new Person(nameBox.Text, addressBox.Text, int.Parse(ageBox.Text), long.Parse(mobileBox.Text), typeComboBox.Text));
         }
 
         private void listBtn_Click(object sender, EventArgs e)
